Reorganize source into server and client directories
