+++ .gitignore
@@ -1,5 +1,3 @@
 coverage
+dist
 node_modules
-
-/features.json
-/public/bundle.*
