+++ __tests__/loader.js
@@ -1,10 +1,10 @@
 const cheerio = require('cheerio');
 const fs = require('fs');
-const scraper = require('../lib/scraper');
+const loader = require('../server/loader');
 const url = require('url');
 
-describe('scraper', () => {
-  const tableHTML = fs.readFileSync(__dirname + '/pages/table.html', 'utf8');
+describe('loader', () => {
+  const tableHTML = fs.readFileSync('__tests__/pages/table.html', 'utf8');
   const $ = cheerio.load(tableHTML);
 
   const expectedTechnologies = ['git', 'mercurial'];
@@ -24,25 +24,25 @@ describe('scraper', () => {
 
   describe('.scrapeTechnologies()', () => {
     it('scrapes all technology names from a table header row', () => {
-      expect(scraper.scrapeTechnologies($('tr').eq(0), $)).toEqual(expectedTechnologies);
+      expect(loader.scrapeTechnologies($('tr').eq(0), $)).toEqual(expectedTechnologies);
     });
   });
 
   describe('.scrapeFeature()', () => {
     it('scrapes a feature from a table row', () => {
-      expect(scraper.scrapeFeature($('tr').eq(1), $, expectedTechnologies)).toEqual(expectedFeatures[0]);
+      expect(loader.scrapeFeature($('tr').eq(1), $, expectedTechnologies)).toEqual(expectedFeatures[0]);
     });
   });
 
   describe('.scrapeTable()', () => {
     describe('with a table with no extra data', () => {
       it('scrapes all features from the table', () => {
-        expect(scraper.scrapeTable($('table'), $)).toEqual(expectedFeatures);
+        expect(loader.scrapeTable($('table'), $)).toEqual(expectedFeatures);
       });
     });
 
     describe('with a table with extra data', () => {
-      const tableHTML = fs.readFileSync(__dirname + '/pages/table_extra.html', 'utf8');
+      const tableHTML = fs.readFileSync('__tests__/pages/table_extra.html', 'utf8');
       const $ = cheerio.load(tableHTML);
 
       const expectedFeatures = [{
@@ -61,14 +61,14 @@ describe('scraper', () => {
       }];
 
       it('scrapes all features from the table, ignoring extra data', () => {
-        expect(scraper.scrapeTable($('table'), $)).toEqual(expectedFeatures);
+        expect(loader.scrapeTable($('table'), $)).toEqual(expectedFeatures);
       });
     });
   });
 
   describe('.getPages()', () => {
     it('gets a list of all Rosetta Stone pages from http://hyperpolyglot.org/', () => {
-      return scraper.getPages().then(pages => {
+      return loader.getPages().then(pages => {
         expect(pages).not.toHaveLength(0)
         pages.forEach(page => {
           var pageURL = url.parse(page);
@@ -83,7 +83,7 @@ describe('scraper', () => {
     jasmine.DEFAULT_TIMEOUT_INTERVAL = 60 * 1000;
 
     it('scrapes features from the Hyperpolyglot website', () => {
-      return expect(scraper.scrape()).resolves.not.toHaveLength(0);
+      return expect(loader.scrape()).resolves.not.toHaveLength(0);
     });
   });
 });
