+++ client/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
     <!-- styles-->
-    <link rel="stylesheet" href="/bundle.css">
+    <link rel="stylesheet" href="/main.css">
   </head>
   <body>
     <!-- header-->
@@ -45,6 +45,6 @@
     </footer>
 
     <!-- scripts-->
-    <script src="/bundle.js"></script>
+    <script src="/main.js"></script>
   </body>
 </html>
